+++ src/main/java/com/bugdb/repository/SeverityRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import com.bugdb.domain.Severity;
 
 
-public interface SeverityRepository extends JpaRepository<Severity, Long>, JpaSpecificationExecutor<Severity> {
+public interface SeverityRepository extends JpaRepository<Severity, Integer>, JpaSpecificationExecutor<Severity> {
 }
